pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "narrate"
version = "0.1.0"
description = "Local text-to-speech with Kokoro and Chatterbox engines"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Manim Studio" },
]
keywords = ["tts", "text-to-speech", "kokoro", "chatterbox", "narration", "voice"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
    "soundfile>=0.12",
    "numpy",
]

[project.optional-dependencies]
kokoro = [
    "kokoro-onnx>=0.4",
    "onnxruntime>=1.17",
]
chatterbox = [
    "chatterbox-tts>=0.1",
    "torch>=2.0",
    "torchaudio>=2.0",
]
play = [
    "sounddevice>=0.4",
]
all = [
    "narrate[kokoro,chatterbox,play]",
]

[project.urls]
Homepage = "https://github.com/user/narrate"
Repository = "https://github.com/user/narrate"

[tool.hatch.build.targets.wheel]
packages = ["src/narrate"]
